requirements.txt
flask
azure-ai-textanalytics
azure-cognitiveservices-speech
moviepy